requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
numpy==1.22.4
opencv-contrib-python==4.6.0.66
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
pyyaml
djangorestframework
